cumulus_test_service
Module bench_utils

Utilities used for benchmarking

Functions

create_benchmark_accounts

Create accounts by deriving from Alice

create_benchmarking_transfer_extrinsics

Creates transfer extrinsics pair-wise from elements of src_accounts to dst_accounts.

extrinsic_set_time

Create a timestamp extrinsic ahead by MinimumPeriod of the last known timestamp

extrinsic_set_validation_data

Create a set validation data extrinsic

get_wasm_module

Prepare cumulus test runtime for execution

import_block

Import block into the given client and make sure the import was successful

set_glutton_parameters

Create a block containing setup extrinsics for the glutton pallet.