Module cumulus_test_service::chain_spec

Structs

Extensions

The extensions for the ChainSpec.

ExtensionsFork

GenesisExt

Extension for the genesis config to add custom keys easily.

Functions

get_account_id_from_seed

Helper function to generate an account ID from seed.

get_chain_spec

Get the chain spec for a specific parachain ID.

get_chain_spec_with_extra_endowed

Get the chain spec for a specific parachain ID. The given accounts are initialized with funds in addition to the default known accounts.

get_from_seed

Helper function to generate a crypto pair from seed

testnet_genesis

Creates a local testnet genesis with endowed accounts.

testnet_genesis_with_default_endowed

Local testnet genesis for testing.

Type Aliases

ChainSpec

Specialized ChainSpec for the normal parachain runtime.