Function cumulus_test_service::chain_spec::get_chain_spec_with_extra_endowed

pub fn get_chain_spec_with_extra_endowed(
    id: ParaId,
    extra_endowed_accounts: Vec<AccountId>
) -> ChainSpec

Get the chain spec for a specific parachain ID. The given accounts are initialized with funds in addition to the default known accounts.