cumulus_test_service

Function start_node_impl

pub async fn start_node_impl<RB, Net: NetworkBackend<NodeBlock, Hash>>(
    parachain_config: Configuration,
    collator_key: Option<CollatorPair>,
    relay_chain_config: Configuration,
    wrap_announce_block: Option<Box<dyn FnOnce(AnnounceBlockFn) -> AnnounceBlockFn>>,
    fail_pov_recovery: bool,
    rpc_ext_builder: RB,
    collator_options: CollatorOptions,
    proof_recording_during_import: bool,
    use_slot_based_collator: bool,
) -> Result<(TaskManager, Arc<Client>, Arc<dyn NetworkService>, RpcHandlers, TransactionPool, Arc<Backend>)>
where
    RB: Fn(Arc<Client>) -> Result<RpcModule<()>, Error> + Send + 'static,

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.