Crate cumulus_test_service

Crate used for testing with Cumulus.

Re-exports

pub use cumulus_test_runtime as runtime;

Structs

Extensions
The extensions for the [ChainSpec].

ExtensionsFork

NativeExecutor
A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

RuntimeExecutor
A unit struct which implements NativeExecutionDispatch feeding in the hard-coded runtime.

TestNode
A Cumulus test node instance used for testing.

TestNodeBuilder
A builder to create a TestNode.

Enums

Keyring
Set of test accounts.

Functions

construct_extrinsic
Construct an extrinsic that can be applied to the test runtime.

get_account_id_from_seed
Helper function to generate an account ID from seed.

get_chain_spec
Get the chain spec for a specific parachain ID.

get_from_seed
Helper function to generate a crypto pair from seed

initial_head_data
Returns the initial head data for a parachain ID.

local_testnet_genesis
Local testnet genesis for testing.

new_partial
Starts a ServiceBuilder for a full service.

node_config
Create a Cumulus Configuration.

run_relay_chain_validator_node
Run a relay-chain validator node.

Type Definitions

AnnounceBlockFn
The signature of the announce block fn.

Client
The client type being used by the test service.