Crate cumulus_test_service

Crate used for testing with Cumulus.

Re-exports

pub use cumulus_test_runtime as runtime;

pub use chain_spec::*;

Modules

chain_spec

Structs

RuntimeExecutor

Native executor instance.

TestNode

A Cumulus test node instance used for testing.

TestNodeBuilder

A builder to create a TestNode.

Enums

Consensus

Keyring

Set of test accounts.

Functions

construct_extrinsic

Construct an extrinsic that can be applied to the test runtime.

fetch_nonce

Fetch account nonce for key pair

initial_head_data

Returns the initial head data for a parachain ID.

new_partial

Starts a ServiceBuilder for a full service.

node_config

Create a Cumulus Configuration.

run_relay_chain_validator_node

Run a relay-chain validator node.

start_node_impl

Start a node with the given parachain Configuration and relay chain Configuration.

Type Definitions

AnnounceBlockFn

The signature of the announce block fn.

Client

The client type being used by the test service.

TransactionPool

Transaction pool type used by the test service