Function cumulus_test_service::start_node_impl

pub async fn start_node_impl<RB>(
    parachain_config: Configuration,
    collator_key: Option<CollatorPair>,
    relay_chain_config: Configuration,
    para_id: ParaId,
    wrap_announce_block: Option<Box<dyn FnOnce(AnnounceBlockFn) -> AnnounceBlockFn>>,
    fail_pov_recovery: bool,
    rpc_ext_builder: RB,
    consensus: Consensus,
    collator_options: CollatorOptions
) -> Result<(TaskManager, Arc<Client>, Arc<NetworkService<Block, H256>>, RpcHandlers, TransactionPool)>where
    RB: Fn(Arc<Client>) -> Result<RpcModule<()>, Error> + Send + 'static,

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.